+++ Module2Assignment/CallingMethodsAssign/Program.cs
@@ -8,12 +8,12 @@ namespace CallingMethodsAssign
         {
             Console.WriteLine("Convert pounds to kilogram.");
             Console.WriteLine("Enter a value in pounds:");
-            decimal lbs = Convert.ToDecimal(Console.ReadLine());
-            decimal inKg = Operation1.Kilogram(lbs);
+            double lbs = Convert.ToDouble(Console.ReadLine());
+            double inKg = Operation1.Kilogram(lbs);
             Console.WriteLine(inKg.ToString());
-            decimal inOz = Operation1.Ounce(lbs);
+            double inOz = Operation1.Ounce(lbs);
             Console.WriteLine(inOz.ToString());
-            decimal inGram = Operation1.Grams(lbs);
+            double inGram = Operation1.Grams(lbs);
             Console.WriteLine(inGram.ToString());
         }
     }
